address merge conflicts and move the cvx tests to different category
